made apikey env var + formatted omdb response
